Move Scales to ES6 module
